Added searching of journal entries.
Each JournalItem row maintains a keywords field which is all the words from all
the entries deduped and combined into a single string. We can then use this to
do an actual SQL query for the JournalItems that match a given text query
instead of loading all thousands of them and querying in memory. The best we
can do with these tools.

Also wrangled with schema migration. OMG, how terrible. But eventually I got
everything working and earned a few battle scars that will guide any future
schema migrations I might be foolish enough to undertake.

+++ InputOutput/Journal/JournalEntryRow.swift
@@ -1,9 +1,31 @@
 import SwiftUI
 
+struct ReadonlyJournalEntryRow: View {
+  var entry :JournalEntry
+  
+  var body: some View {
+    HStack {
+      Image(systemName: "circle.fill").resizable().frame(width: 10, height: 10)
+      Text(entry.text)
+      
+      Spacer()
+      if let tags = entry.tags {
+        ForEach(tags, id: \.self) { tag in
+          Text(tag).foregroundColor(.white)
+            .padding([.leading, .trailing], 10)
+            .padding([.top, .bottom], 5)
+            .background(.gray).cornerRadius(15)
+        }
+      }
+    }
+    .frame(minHeight: 30)
+  }
+}
+
 struct JournalEntryRow: View {
   @Binding var entry: JournalEntry
   @State var isEditing: Bool
-  var onDelete :(UUID) -> Void
+  var onDelete :((UUID) -> Void)?
   var onEditingDone :() -> Void
   @State private var restoreText = ""
   @FocusState private var isFocused: Bool
@@ -30,9 +52,11 @@ struct JournalEntryRow: View {
           })
         #endif
         #if os(macOS)
-        Button(action: { onDelete(entry.id) }) {
-          Image(systemName: "trash")
-        }.buttonStyle(PlainButtonStyle())
+        if let onDelete = onDelete {
+          Button(action: { onDelete(entry.id) }) {
+            Image(systemName: "trash")
+          }.buttonStyle(PlainButtonStyle())
+        }
         #endif
       } else {
         Text(entry.text)
